Move #core Skills out of DEFAULT and into each DEFAULT.platform file,… (#452)
With this change, the definition of the DEFAULT-SKILLS files changes.  Previously it was assumed that MSM would install ```DEFAULT-SKILLS``` + ```DEFAULT-SKILLS.<platform>``` (if the platform file exists).  Now MSM can install DEFAULT-SKILLS.<platform> _or_ (if it doesn't exist) ```DEFAULT-SKILLS```.

The change to MSM has not occurred yet, but the default files had to be updated before the change began.  This resolves #412 

Details:
* Remove stocks and weather Skills from KDE default
* Remove mycroft-wiki Skill from KDE default
* Update DEFAULT-SKILLS.kde
  - add mycroft-playback-control
  - add mycroft-speak
  - add mycroft-volume
* Update DEFAULT-SKILLS.mycroft_mark_1
  - add mycroft-playback-control
  - add mycroft-speak
  - add mycroft-volume
* Update DEFAULT-SKILLS.picroft
  - add mycroft-playback-control
  - add mycroft-speak
  - add mycroft-volume