Use unique paths for log temp file storage
This extends the change in #1516.

When transferring the Mycroft logs to the report host, the zip
file was being written to a common directory. In the event that
multiple jobs were writing to or reading from this directory at
the same time, conflicts could occur.

This ensures that the mycroft-logs.zip file is written to a unique
path and cleaned up afterward.