Remove Skill author from test definition
The Skill author was added to improve matching of Skills
however this is not yet supported by mycroft-core and
has actually reduced the confidence level for matches.

The fix in #1520 did not account for the submodule path
being used, which is inherently unique. So it is not
possible for two Skills to have the same submodule path.
As such the fix to prevent characters being accidentally
trimmed off the skill_submodule_path should be sufficient
to have all Skills match exactly.